Created script that spawns power-up in a random location. Made a power-up prefab. TO DO:In a power-up script Make power-up increase player speed Make power-up disappear on contact Make power up falls slower
